+++ packages/inference/src/tasks/custom/request.ts
@@ -1,5 +1,5 @@
-import type { Options, RequestArgs } from "../types";
-import { makeRequestOptions } from "../lib/makeRequestOptions";
+import type { Options, RequestArgs } from "../../types";
+import { makeRequestOptions } from "../../lib/makeRequestOptions";
 
 /**
  * Primitive to make custom calls to the inference API
