+++ packages/inference/src/tasks/cv/imageClassification.ts
@@ -1,5 +1,5 @@
-import type { ImageClassificationArgs, ImageClassificationReturn, Options } from "../types";
-import { request } from "./request";
+import type { ImageClassificationArgs, ImageClassificationReturn, Options } from "../../types";
+import { request } from "../custom/request";
 
 /**
  * This task reads some image input and outputs the likelihood of classes.
