+++ packages/inference/src/tasks/index.ts
@@ -1,22 +1,22 @@
-export { audioClassification } from "./audioClassification";
-export { automaticSpeechRecognition } from "./automaticSpeechRecognition";
-export { conversational } from "./conversational";
-export { featureExtraction } from "./featureExtraction";
-export { fillMask } from "./fillMask";
-export { imageClassification } from "./imageClassification";
-export { imageSegmentation } from "./imageSegmentation";
-export { imageToText } from "./imageToText";
-export { objectDetection } from "./objectDetection";
-export { questionAnswering } from "./questionAnswering";
-export { request } from "./request";
-export { sentenceSimilarity } from "./sentenceSimilarity";
-export { streamingRequest } from "./streamingRequest";
-export { summarization } from "./summarization";
-export { tableQuestionAnswering } from "./tableQuestionAnswering";
-export { textClassification } from "./textClassification";
-export { textGeneration } from "./textGeneration";
-export { textGenerationStream } from "./textGenerationStream";
-export { textToImage } from "./textToImage";
-export { tokenClassification } from "./tokenClassification";
-export { translation } from "./translation";
-export { zeroShotClassification } from "./zeroShotClassification";
+export { audioClassification } from "./audio/audioClassification";
+export { automaticSpeechRecognition } from "./audio/automaticSpeechRecognition";
+export { conversational } from "./nlp/conversational";
+export { featureExtraction } from "./nlp/featureExtraction";
+export { fillMask } from "./nlp/fillMask";
+export { imageClassification } from "./cv/imageClassification";
+export { imageSegmentation } from "./cv/imageSegmentation";
+export { imageToText } from "./cv/imageToText";
+export { objectDetection } from "./cv/objectDetection";
+export { questionAnswering } from "./nlp/questionAnswering";
+export { request } from "./custom/request";
+export { sentenceSimilarity } from "./nlp/sentenceSimilarity";
+export { streamingRequest } from "./custom/streamingRequest";
+export { summarization } from "./nlp/summarization";
+export { tableQuestionAnswering } from "./nlp/tableQuestionAnswering";
+export { textClassification } from "./nlp/textClassification";
+export { textGeneration } from "./nlp/textGeneration";
+export { textGenerationStream } from "./nlp/textGenerationStream";
+export { textToImage } from "./cv/textToImage";
+export { tokenClassification } from "./nlp/tokenClassification";
+export { translation } from "./nlp/translation";
+export { zeroShotClassification } from "./nlp/zeroShotClassification";
