+++ packages/inference/src/tasks/nlp/summarization.ts
@@ -1,5 +1,5 @@
-import type { Options, SummarizationArgs, SummarizationReturn } from "../types";
-import { request } from "./request";
+import type { Options, SummarizationArgs, SummarizationReturn } from "../../types";
+import { request } from "../custom/request";
 
 /**
  * This task is well known to summarize longer text into shorter text. Be careful, some models have a maximum length of input. That means that the summary cannot handle full books for instance. Be careful when choosing your model.
